refactored manager tests into model tests
